LayoutTests/platform/chromium-linux/plugins/keyboard-events-expected.txt

CONSOLE MESSAGE: line 18: PLUGIN: getFocusEvent
CONSOLE MESSAGE: line 18: PLUGIN: mouseDown at (12, 12)
CONSOLE MESSAGE: line 19: PLUGIN: mouseUp at (12, 12)
CONSOLE MESSAGE: line 22: PLUGIN: keyDown ' '
CONSOLE MESSAGE: line 22: PLUGIN: keyUp ' '
CONSOLE MESSAGE: line 23: PLUGIN: keyDown ' '
CONSOLE MESSAGE: line 23: PLUGIN: keyUp ' '
CONSOLE MESSAGE: line 24: PLUGIN: keyDown ' '
CONSOLE MESSAGE: line 24: PLUGIN: keyUp ' '

This test checks if a plug-in can receive keyboard events sent from eventSender. This is a test for Bug 34936.
